+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Show CREATED AT instead of AGE in deploy list
+- Delete configmap if a deploy remove the nginx
 
 ### Added
 - Share application directory with nginx sidecar
@@ -21,7 +22,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Update `gorm` version to v1.9.1
 - Update `go-sql-driver/mysql`
-- Delete configmap if a deploy remove the nginx
 
 ### Fixed
 - `.teresaignore` behavior (to work like `.gitignore`)